+++ src/HuffmanCodebook.h
@@ -7,11 +7,11 @@ using namespace std;
 class HuffmanCodebook
 {
 public:
-	uint32_t size;
+	uint16_t size;
 	uint32_t* codes;
 	uint8_t* codes_length;
 	uint8_t* codes_value;
 	
-	HuffmanCodebook(uint32_t size);
+	HuffmanCodebook(uint16_t size);
 	~HuffmanCodebook();
 };
